Package as installable video-use CLI tool
- src/ layout package with console entry point (video-use transcribe|pack|timeline|grade|render)
- video-use skill / where / key / doctor for agent-driven install
- SKILL.md + manim-video sub-skill bundled as package data
- drop unused librosa + matplotlib hard deps (classic install failure)
- API key resolution: env -> ./.env -> ~/.config/video-use/.env -> repo .env
- UTF-8 stdout forcing for Windows consoles
- helpers/*.py kept as back-compat shims for clone+symlink installs
- install.md rewritten as uv tool install fast path; uv.lock added
