+++ src/video_use/__init__.py
@@ -1,0 +1,3 @@
+"""video-use — edit videos with coding agents."""
+
+__version__ = "0.2.0"
