+++ src/video_use/skills/manim-video/references/animation-design-thinking.md
