+++ src/video_use/skills/manim-video/references/animations.md
