+++ src/video_use/skills/manim-video/references/camera-and-3d.md
