+++ src/video_use/skills/manim-video/references/equations.md
