+++ src/video_use/skills/manim-video/references/paper-explainer.md
