+++ src/video_use/skills/manim-video/references/rendering.md
