+++ src/video_use/skills/manim-video/references/scene-planning.md
