+++ src/video_use/skills/manim-video/references/visual-design.md
