+++ src/video_use/skills/manim-video/scripts/setup.sh
